Altered keys stuff to deal with the new format JSON/XML.
Test coverage almost back up to 100%, but there's a line I haven't
figured out how to it yet, or rather looking at the test is appears as
if it should be being reached, but according to rcov isn't.

Hell of a lot of moving parts in pachube_data_formats, so I hope I've
got all the bits.