Refactor
Was kind of weird trying to deal with Scope and Resource objects when
they were some kind of inner classes to the parent Key object - by weird
I mean that it made it hard to deal with them when dealing with AR
objects in rails, so actually it would be a heck of a lot easier if they
existed as full PDF objects in their own right, with corresponding real
AM objects in Pachweb.

Also renamed Scope to Permission as the term Scope isn't the best (I
think).